A set of minor cleanups/warnings/bug-fixes. (Thanks to Albert Kwon who mentioned many of these from Ubuntu/Debian) tools/cdk/export/lib/hdl: Remove built files from tree that were incorrectly committed core/container/ctests/bin/src/RCCLbTest_main.cxx: warnings core/container/ctests/workers/include/LoopbackWorker.h: warnings core/container/ctests/workers/src/CosumerWorker.c: warnings core/container/ctests/workers/src/LoopbackWorker.c: warnings core/container/hdl_container/bin/src/testRpl_main.cxx: missing #include <string.h> core/container/hdl_container/impl/include/PciScanner.h: warnings core/container/hdl_container/impl/src/PciScanner.cxx: warnings core/container/hdl_container/impl/src/XyzContainer.cxx: missing #include <limits.h> core/container/interfaces/api/include/OcpiContainerApi.h: missing #include <string> core/container/interfaces/api/src/OcpiContainerInterface.cxx: warnings core/container/ocl_container/impl/src/OclContainer.cxx: missing #include <limits.h> core/container/rcc_container/src/OcpiRccPort.cxx: added FIXME for API error check core/dataplane/rdma_drivers/sockets/src/DtSocketsXfer.cxx: warnings core/dataplane/rdma_utils/utils/src/xfer.cxx: warnings core/dataplane/transport/impl/src/OcpiTransport.cxx: warnings core/library/include/OcpiLibraryManager.h: warnings core/library/src/OcpiComponentLibrary.cxx: missing #include <stdlib.h> core/library/src/OcpiLibraryManager.cxx: warnings core/local/util/fasttime/include/config.h: warnings core/local/util/parentChild/include/OcpiParentChild.h: warnings core/local/util/parentChild/src/OcpiParentChild.cxx: missing #include <stdint.h> do_omniidl.csh: remove tsch dependency to csh always works ($#->$#argv) test/hdl/src/nft.c: use sched_yield rather than pthread_yield for portability test/rcc/rot/rot.c: warnings tools/cdk/include/util.mk: support directory paths with embedded spaces tools/cdk/ocpigen/ocpigen/src/wip_parse.cxx: fix pathname merging when path has slashes

+++ core/container/ctests/workers/src/CosumerWorker.c
@@ -109,11 +109,11 @@ static RCCResult ConsumerWorker_run(RCCWorker *this_,RCCBoolean timedout,RCCBool
 {
   ( void ) timedout;
   ( void ) newRunCondition;
-  int ncount, *b;
+  uint32_t ncount, *b;
   uint32_t len,n;
   ConsumerWorkerStaticMemory *mem = this_->memories[0];
   ConsumerWorkerProperties *props = this_->properties;
-  int passed = 1;
+  unsigned passed = 1;
 
 #ifdef TIME_TP
   double          usecs;
@@ -152,7 +152,7 @@ static RCCResult ConsumerWorker_run(RCCWorker *this_,RCCBoolean timedout,RCCBool
   
 #define CHECK_DATA
 #ifdef CHECK_DATA
-  b = (int*)(in_buffer);
+  b = (unsigned*)(in_buffer);
 
 #define RESYNC
 #ifdef RESYNC
@@ -166,7 +166,7 @@ static RCCResult ConsumerWorker_run(RCCWorker *this_,RCCBoolean timedout,RCCBool
   */
 
 
-  if ( *b != (int)mem->b_count ) {
+  if ( *b != mem->b_count ) {
     printf("ERROR!! Dropped a buffer, got buffer %d, expected %d\n", 
            *b, mem->b_count );
     dropped_b++;
@@ -178,7 +178,7 @@ static RCCResult ConsumerWorker_run(RCCWorker *this_,RCCBoolean timedout,RCCBool
   ncount = 0;
   for (n=4; n<len+4; n++) {
     if ( (in_buffer[n] != (char)(n+mem->b_count)%23) && (ncount++ < 100000) ) {
-      printf("Consumer(%lu, %lu, b-> %d): Data integrity error(%d) !!, expected %d, got %d\n", 
+      printf("Consumer(%u, %u, b-> %d): Data integrity error(%d) !!, expected %d, got %d\n", 
              props->startIndex, len, mem->b_count,n, (char)(n+mem->b_count)%23, in_buffer[n]);
       passed = 0;
     }
